Remove patterns from word strings
Patterns such as 'vpm' (vowel pattern marker) should not be included
in word strings meta_word, paradigmatic_form and surface_form, since
they are not actually morphemes.
TODO: now "morphemes" with type 'vpm' are filtered. But the patterns
are defined in the word_grammar, and their names should not be
hardcoded. They should instead be filtered by their position in the
word_grammar (mt.pos == 4).